feat: move k8s_object_handler to k8s_watcher __init__
